[http_request] Add support for chunked HTTP responses
This change removes the unnecessary trimming of buffer length, which
enables reading responses without knowing the actual length, first
(chunked responses).

When providing a buffer to read from a stream it is only necessary to
specify the length of the buffer. The actual amount of available data to
read is usually known by the stream. It won't keep waiting for more data
when the stream had already been closed or all data were received.